7-Zip Version 21.02
What's new after 7-Zip 20.02 alpha:
https://sourceforge.net/p/sevenzip/discussion/45797/thread/dd4edab390/

What's new after 7-Zip 21.00 alpha:
https://sourceforge.net/p/sevenzip/discussion/45797/thread/d401ab2966/

What's new after 7-Zip 21.01 alpha:
https://sourceforge.net/p/sevenzip/discussion/45797/thread/7fe6c21efa/

+++ CPP/7zip/Archive/Tar/TarOut.cpp
@@ -62,7 +62,7 @@ static void WriteOctal_12_Signed(char *s, Int64 val)
 {
   if (val >= 0)
   {
-    WriteOctal_12(s, val);
+    WriteOctal_12(s, (UInt64)val);
     return;
   }
   s[0] = s[1] = s[2] = s[3] = (char)(Byte)0xFF;
